+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.7.2] - 2018-12-31
+### Changed
+- Use `term_size` instead of `termion`
+- Color tweaks
+
 ## [0.7.1] - 2018-12-31
 ### Changed
 - Merged `ip_addr -> {general, specific}` into `ip_addr -> information`